% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph.R
\name{bar_graph}
\alias{bar_graph}
\title{Create a bar graph for a given variable}
\usage{
bar_graph(df, x, y, xlab, ylab, fill_lab, title)
}
\arguments{
\item{x}{The name of the discrete variable that will be on the x-axis of the bar graph.}

\item{y}{The name of the numerical variable that will be on the y-axis of the bar graph.}

\item{xlab}{The label of the x-axis.}

\item{ylab}{The label of the y-xis.}

\item{fill_lab}{The label of the fill variable.}

\item{title}{The title of the bar graph.}

\item{data}{The dataset that contains the x variable.}
}
\value{
A bar graph for given discrete variable, x, and numerical variable y
}
\description{
This function takes in a dataset, a variable x from the dataset, a variable y from the dataset, an x-axis label, a y-axis label, a fill label, and a title for the plot
}
\examples{
# bar_graph(iris, iris$Species, iris$Sepal.Width, "Species", "Sepal Width", "Species", "Iris Graph")



}
