% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_plot.R
\name{hist_plot}
\alias{hist_plot}
\title{Histogram Plot
Create a histogram for a given variable}
\usage{
hist_plot(data, var, xlab, ylab, title, text_size)
}
\arguments{
\item{data}{The dataset that contains the x variable.}

\item{var}{The name of the numerical variable that will be on the x-axis of the histogram.}

\item{xlab}{The label of the x-axis.}

\item{ylab}{The label of the y-xis.}

\item{title}{The title of the histogram plot.}

\item{text_size}{The size of the plot's text.}
}
\value{
A histogram plot for given numerical variable
}
\description{
This function takes in a dataset, a variable x from the dataset, an x-axis label, a y-axis label, a title for the plot and the text size.
}
\examples{
# hist_plot(mtcars, mpg, "Miles Per Gallon", "Count", "Histogram for MPG of Cars", 10)

}
