% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{linearmodel}
\alias{linearmodel}
\title{Linear Regression Model
Create a linear regression model}
\usage{
linearmodel(recipe, data)
}
\arguments{
\item{recipe}{The recipe for the model.}

\item{data}{The dataset for the model.}
}
\value{
A list that includes the intercept and coefficients for the linear model.
}
\description{
This function takes in a recipe and a data frame.
}
\examples{
linearmodel(recipe(mpg ~ hp + wt, mtcars), mtcars)

}
