% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rmspe.R
\name{model_rmspe}
\alias{model_rmspe}
\title{Model RMSPE}
\usage{
model_rmspe(model, testing_data, var)
}
\arguments{
\item{model}{The linear or k-nn regression model}

\item{testing_data}{The testing dataset for the model}

\item{var}{The name (as a string) of the column for the response variable}
}
\value{
RMSPE value as a number
}
\description{
Output the RMPSE value of a linear or k-nn regression model
}
\details{
This function takes in a model, a data frame and a column name as a string
}
\examples{
split <- rsample::initial_split(iris)
iris_train <- rsample::training(split)
iris_test <- rsample::testing(split)
iris_model <- linearmodel(recipes::recipe(Sepal.Length ~ Sepal.Width, iris_train), iris_train)
model_rmspe(iris_model, iris_test, "Sepal.Length")

}
